refactor: placing files in an appropriate place and changing logic in icons page
