--- apps/docs/src/components/Icon.astro
@@ -1,7 +1,0 @@
----
-import { Icons } from '@selleo/core/src/Icon';
-import Preview from './Preview.astro';
-import CodeWithCopy from './CodeWithCopy.astro';
----
-
-<Icons client:load />
