+++ apps/docs/src/components/IconPage.astro
@@ -1,0 +1,5 @@
+---
+import { Icons } from './Icons/Icons';
+---
+
+<Icons client:load />
